services/microsoft-apis/public/requirements.txt
aiohttp
aiomysql
cryptography
fastapi[standard]==0.115.0
msal
python-dotenv
pydantic==2.9.2
sqlalchemy==2.0.36
PyOpenSSL
pyjwt
aio-pika
asyncache